Solved issue with the nested routes for the categories and implemented the product display
